requirements.txt
torch
torchvision
numpy
matplotlib
pandas
jupyter
scikit-learn
scipy
Pillow